pyproject.toml
[tool.poetry]
name = "chatbot-langchain"
version = "0.1.0"
description = "Chatbot: OpenAI + LangChain"
authors = ["priordd <rodrigo.prior@datadoghq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.4"
openai = "^1.52.2"
python-dotenv = "^1.0.1"
chainlit = "^1.3.1"
langchain-openai = "^0.2.3"
ddtrace = "^2.14.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
